[ES-1614194] Include OAuth HTML resources (#1107)
## Description
<!-- Provide a brief summary of the changes made and the issue they aim
to address.-->
OAuth U2M is broken in databricks-jdbc for AWS and GCP. This PR fixes it
by including the required resources needed by SDK classes.

## Testing
<!-- Describe how the changes have been tested-->
Tested by locally installing the jar and running the reproducer on the
ticket.

## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->